# docs/swagger.yaml
basePath: /
definitions:
...
  handler.RespBody:
    properties:
      code:
        description: http code
        type: integer
      data:
        description: response data
      msg:
        description: response message
        type: string
      reason:
        description: reason key
        type: string
    type: object
...
  schema.BatchRemoveProjectReq:
    properties:
      ids:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - ids
    type: object
...
  schema.CreateProjectReq:
    properties:
      code_type:
        enum:
        - 1
        - 2
        - 3
        type: integer
      content:
        minLength: 1
        type: string
      cover:
        type: string
      description:
        maxLength: 1000
        minLength: 1
        type: string
      question_id:
        type: string
      repo_url:
        additionalProperties:
          type: string
        type: object
      tags:
        items:
          type: string
        type: array
      title:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - code_type
    - content
    - description
    - question_id
    - repo_url
    - title
    type: object
...
  schema.ProjectPageResp:
    properties:
      count:
        type: integer
      list:
        items:
          $ref: '#/definitions/schema.ProjectResp'
        type: array
    type: object
  schema.ProjectResp:
    properties:
      answer_count:
        type: integer
      code_type:
        type: integer
      code_type_name:
        type: string
      content:
        type: string
      cover:
        type: string
      created_at:
        type: integer
      description:
        type: string
      id:
        type: string
      question_id:
        type: string
      repo_url:
        additionalProperties:
          type: string
        type: object
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: integer
      user_id:
        type: string
      user_info:
        $ref: '#/definitions/schema.UserBasicInfo'
      view_count:
        type: integer
      vote_count:
        type: integer
    type: object
...
  schema.RemoveProjectReq:
    properties:
      id:
        type: string
    required:
    - id
    type: object
...
  schema.UpdateProjectReq:
    properties:
      code_type:
        enum:
        - 1
        - 2
        - 3
        type: integer
      content:
        minLength: 1
        type: string
      cover:
        type: string
      description:
        maxLength: 1000
        minLength: 1
        type: string
      id:
        type: string
      repo_url:
        additionalProperties:
          type: string
        type: object
      tags:
        items:
          type: string
        type: array
      title:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - code_type
    - content
    - description
    - id
    - repo_url
    - title
    type: object
...
  schema.UserBasicInfo:
    properties:
      avatar:
        type: string
      display_name:
        type: string
      id:
        type: string
      language:
        type: string
      location:
        type: string
      rank:
        type: integer
      status:
        type: string
      username:
        type: string
      website:
        type: string
    type: object
...
info:
  contact: {}
  description: Apache Answer API
  title: Apache Answer
paths:
...
  /answer/api/v1/project:
    delete:
      consumes:
      - application/json
      description: 删除项目
      parameters:
      - description: project
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.RemoveProjectReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: 删除项目
      tags:
      - Project
    post:
      consumes:
      - application/json
      description: 添加项目
      parameters:
      - description: project
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.CreateProjectReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: 添加项目
      tags:
      - Project
    put:
      consumes:
      - application/json
      description: 更新项目
      parameters:
      - description: project
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.UpdateProjectReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: 更新项目
      tags:
      - Project
  /answer/api/v1/project/batch-delete:
    delete:
      consumes:
      - application/json
      description: 批量删除项目
      parameters:
      - description: project ids
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.BatchRemoveProjectReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: 批量删除项目
      tags:
      - Project
  /answer/api/v1/project/info:
    get:
      consumes:
      - application/json
      description: 获取项目详情
      parameters:
      - description: project id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.ProjectResp'
              type: object
      summary: 获取项目详情
      tags:
      - Project
  /answer/api/v1/project/page:
    get:
      consumes:
      - application/json
      description: 获取项目分页列表
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: page_size
        type: integer
      - description: search
        in: query
        name: search
        type: string
      - description: code type
        in: query
        name: code_type
        type: integer
      - description: tags
        in: query
        name: tags
        type: string
      - description: order
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.ProjectPageResp'
              type: object
      summary: 获取项目分页列表
      tags:
      - Project
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"